+++ client/csharp_example/Test/HoloGainExample.cs
@@ -4,7 +4,7 @@
  * Created Date: 25/08/2019
  * Author: Shun Suzuki
  * -----
- * Last Modified: 05/09/2019
+ * Last Modified: 20/02/2020
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2019 Hapis Lab. All rights reserved.
@@ -22,23 +22,23 @@ namespace AUTD3SharpTest.Test
         {
             Console.WriteLine("Start HoloGain Test");
 
-            float x = 83.0f;
-            float y = 66.0f;
-            float z = 150.0f;
+            double x = 83.0;
+            double y = 66.0;
+            double z = 150.0;
 
             using (AUTD autd = new AUTD())
             {
                 autd.Open();
-                autd.AddDevice(Vector3f.Zero, Vector3f.Zero);
+                autd.AddDevice(Vector3d.Zero, Vector3d.Zero);
                 autd.AppendModulationSync(AUTD.SineModulation(150)); // AM sin 150 HZ
 
-                Vector3f[] focuses = new[] {
-                    new Vector3f(x - 30, y ,z),
-                    new Vector3f(x + 30, y ,z)
+                Vector3d[] focuses = new[] {
+                    new Vector3d(x - 30, y ,z),
+                    new Vector3d(x + 30, y ,z)
                 };
-                float[] amps = new[] {
-                    1.0f,
-                    1.0f
+                double[] amps = new[] {
+                    1.0,
+                    1.0
                 };
                 autd.AppendGainSync(AUTD.HoloGain(focuses, amps));
